fix: bigint형 -> int형 변환, 관련 파일 변경

+++ backend/analysis-worker/src/main/java/com/ssafy/modera/worker/domain/analysis/repository/AnalysisJobRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.modera.worker.domain.analysis.repository;
 import com.ssafy.modera.worker.domain.analysis.entity.AnalysisJob;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnalysisJobRepository extends JpaRepository<AnalysisJob, Long> {
+public interface AnalysisJobRepository extends JpaRepository<AnalysisJob, Integer> {
 }
